requirements.txt
certifi==2021.5.30
click==8.0.1
dnspython==2.1.0
Flask==2.0.1
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
importlib-metadata==4.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
pymongo==3.11.4
python-dotenv==0.18.0
SQLAlchemy==1.4.19
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.4.1
